+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   to set HTML attributes and a CSS class for filter cells (@vjik)
 - New #362: Add `GridView::captionAttributes()` method and `$attributes` parameter to `GridView::caption()` to set
   HTML attributes for the `caption` tag (@vjik)
+- New #363: Add `rowHeader` parameter to `DataColumn` to render body cells as `<th scope="row">` (@vjik)
 - Enh #358: Make dependency container in `GridView` constructor optional (@vjik)
 - Enh #363: Add `scope="col"` attribute to `GridView` header cells by default (@vjik)